requirements.txt
streamlit
pdfplumber
langchain
langchain_community
langchain_google_genai
python-dotenv
google-generativeai
faiss-cpu
deepeval
guardrails-ai
pydantic

